lib/gstreamer: share common util functions
