+++ lib/gstreamer/util.py
@@ -1,0 +1,16 @@
+###
+### Copyright (C) 2021 Intel Corporation
+###
+### SPDX-License-Identifier: BSD-3-Clause
+###
+
+from ...lib.common import memoize, try_call
+
+@memoize
+def have_gst():
+  return try_call("which gst-launch-1.0") and try_call("which gst-inspect-1.0")
+
+@memoize
+def have_gst_element(element):
+  result = try_call("gst-inspect-1.0 {}".format(element))
+  return result, element
