+++ lib/gstreamer/vaapi/util.py
@@ -4,17 +4,9 @@
 ### SPDX-License-Identifier: BSD-3-Clause
 ###
 
-from ....lib.common import memoize, try_call
+from ....lib.common import memoize
 from ....lib.formats import match_best_format
-
-@memoize
-def have_gst():
-  return try_call("which gst-launch-1.0") and try_call("which gst-inspect-1.0")
-
-@memoize
-def have_gst_element(element):
-  result = try_call("gst-inspect-1.0 {}".format(element))
-  return result, element
+from ....lib.gstreamer.util import *
 
 def get_supported_format_map():
   #The map first entry is for gst element properties;the second entry is for gst caps filters
